+++ generators/app/templates/src/index.jsx
@@ -23,7 +23,7 @@ class App extends React.PureComponent {
   render() {
     const { text } = this.state;
     return (
-      <FlexBox padding="small" width="400px">
+      <FlexBox style={{padding: 'small', width: '400px'}}>
         <Button onClick={this.onClick} text={text} />
       </FlexBox>
     );